Moving `Simulation` and `SimulationResults` to new `simulation` module
