+++ docs/source/pulser.rst
@@ -63,7 +63,7 @@ The :class:`Simulation` class is responsible for the classical simulation of
 a :class:`Sequence`.
 
 
-.. automodule:: pulser.simulation
+.. automodule:: pulser.simulation.simulation
    :members:
    :undoc-members:
 
@@ -72,7 +72,7 @@ Simulation Results
 The :class:`SimulationResults` features methods for the calculation of
 expectation values or emulation of the sampling performed on physical devices.
 
-.. automodule:: pulser.simresults
+.. automodule:: pulser.simulation.simresults
    :members:
    :undoc-members:
